MySQL Stitcher Integration
Summary:
Integrates MySQL Stitcher with the connection tracker model.

Add a SocketTraceConnector Test that tests MySQL Stmt prepare and execute end2end.

Test Plan: Passes newly added test.

Reviewers: #engineering, oazizi, yzhao, zasgar

Reviewed By: #engineering, oazizi

Differential Revision: https://phab.corp.pixielabs.ai/D1459

GitOrigin-RevId: e064b068364200c23e21fbe9ba08c2fd059fa79a

+++ src/stirling/mysql/test_utils.h
@@ -11,7 +11,7 @@ namespace stirling {
 namespace mysql {
 namespace testutils {
 
-std::string GenRawPacket(int packet_num, const std::string& msg);
+std::string GenRawPacket(uint8_t packet_num, const std::string& msg);
 
 std::string GenRequest(ConstStrView command, const std::string& msg);
